Make directory structure the same as the Cards Against Humanity plugin
