+++ plugin_code/dublin_bus/setup.js
@@ -1,4 +1,4 @@
-var DublinBusInfo = require('./controllers/dublin_bus_info.js');
+var DublinBusInfo = require('./app/controllers/dublin_bus_info.js');
 
 module.exports = function(app) {
     var dublinBusInfo = new DublinBusInfo();
